Fix minor issues in site menu
Adds a footer to the main index page. This was done to create a unified
styling for the website.

Moves the page "JDO v JPA : ORM" to the section "General" (instead of
"Metadata") so that it matches its entry in the menu bar.

Fixes the highlighted text for the menu bar entry "API Usage".

Removes spaces from the names of the menu bar entries to avoid
inconsistent spacing when highlighting or not highlighting the name.

Adjusts the search box to search on the site "db.apache.org/jdo".

Adjusts the styling of the footer. Changes the color to white and the
background color to gray. Centers the links in the footer.