Rework README
Restructures the readme to make it easier to read and find the necessary
information.

Drops the instructions on how to use git. If the potential contributor
does not know how to use it, this readme is not the place for them to
learn it.

Removes the entry on committing the generated site artifacts. This
should be done separately to keep the pull requests clean.

Adds links to GitHub resources explaining the basic GitHub workflow.