Add GitHub workflow to auto-deploy site (#6)
Adds a GitHub workflow to auto-deploy the site if any of its sources
have changed and the build was successful.

The action is only run if
- the correct branches are pushed (currently only 'master')
AND
- at least one file in the correct paths was modified (currently
  'src/main/asciidoc/**' and 'src/main/template/**').

If the build was successful, the action copies the build results from
'target/site/*' to 'docs/' and creates a new commit with all changed
files in 'docs/'. This commit is automatically pushed to the branch that
triggered the GitHub action. The new commit is attributed to the author
of the latest commit on that branch.

Removed the note to move the build results into the 'docs/' folder when
creating a commit from the README as this is now done automatically.