Update contribution guidelines (#7)
* Rework README

Restructures the readme to make it easier to read and find the necessary
information.

Drops the instructions on how to use git. If the potential contributor
does not know how to use it, this readme is not the place for them to
learn it.

Removes the entry on committing the generated site artifacts. This
should be done separately to keep the pull requests clean.

Adds links to GitHub resources explaining the basic GitHub workflow.

* Add CONTRIBUTING page

This page will be linked to whenever a new PR is created. It just points
the user to the README and the "Getting Involved" page on the website.

* Update "Get Involved" page

Restructures the sections to make information easier to find.

Adds more information on how to contribute using GitHub.

Adds a section about the need for the Apache Contributor License
Agreement.

Links back to the "Get Involved" page from the "Source Code" page.