Implement rudimentary push logic for deploy action
Replaces the external action with bare-bones logic accomplishing the
same goal. This was done to avoid the usage of external GitHub actions.

This new logic is most likely more brittle than the functionality
provided by the external action. It will fail if the author data of the
HEAD commit can not be read and will also always push the new commit to
the currently checked out branch, i.e. the branch that triggered the
action. The workflow is configured to currently only run on the master
branch, but this behavior should still be kept in mind in case of future
changes to the setup.