Move site build artifacts to separate site branch
Adjusts the deploy action to create a new orphan branch 'publish'
containing only the site build artifacts and site configuration files in
the root directory. This branch is then force-pushed to the repo,
replacing the previous site build.

This approach avoids having to include the site artifacts in the master
branch, reducing the repository size and avoiding including "deploy"
commits in the main history.

Creates a new '.asf.yaml.publish' that will be used to publish the site
branch. Moves the publishing configuration from the main configuration
file to the publishing configuration file. Changes the branch id to
'publish'. Drops the 'staging' configuration as it reportedly does not
function correctly. Furthermore, the current content of the 'staging'
branch is no longer deployable with the new setup as it still uses the
old site deploy scheme.

Removes the '.htaccess' configuration as the redirection is no longer
necessary now that the resources are hosted in the root directory.