Set up branch protection for the 'publish' branch
Enables branch protection for the 'publish' branch used to publish the
site build artifacts. This ensures that the publishing branch is never
accidentally deleted, which could break the currently deployed website.

Force-pushes are explicitly enabled with the set-up branch protection as
new versions of the site are published by force-pushing to the 'publish'
branch.