Add recovery workflow to recreate the site branch
Adds a workflow to recreate the site branch in case it ever gets deleted
on accident or becomes corrupted.

The new workflow to automatically deploy the site relies on there being
a functioning site branch. This workflow offers an recovery option to
recreate a stable version of the site branch with which the deploy
workflow can function.

The workflow does not listen for any events on the repository. It can
only be triggered manually. To do so, go to the "Actions" section of the
GitHub page for the repository and select the action from the menu. At
the top of the list of past workflow runs is a blue bar offering the
option to "Run workflow". Select the option and press the green button
"Run workflow".

When running the workflow, the default selection to run it on the master
branch should not be changed. Even though the workflow is explicitly
configured to check out the master branch, running the action on a
different branch might still have other unintended side effects.