Split up publishing stage of deploy action into multiple steps
Splits up the publishing stage of the deploy action into multiple steps.
This makes it easier to determine where an issue occurred during
debugging. Furthermore, it makes the logs easier to digest as they are
automatically grouped by steps.

Also introduces the multi-line syntax for the site build to make the
used syntax more uniform.