Split up publishing stage of recreate action into multiple steps
Splits up the publishing stage of the action to recreate the publishing
branch into multiple steps. This makes it easier to determine where an
issue occurred during debugging. Furthermore, it makes the logs easier
to digest as they are automatically grouped by steps.

Makes the steps moving the site build results to the correct location
verbose to make debugging easier.

Also introduces the multi-line syntax for the site build to make the
used syntax more uniform.