Set notification settings in .asf.yaml
Adds specific notification settings in the asf.yaml configuration.

Diverts messages regarding new commits being pushed to the 'jdo-commits'
mailing list to avoid clogging up 'jdo-dev'. The notifications for
pull-requests and issues still go to 'jdo-dev'.

Adds the preexisting jira linking options to the configurations. With
the current configuration, the label 'pull-request-available' and a link
to the pull-request/issue is automatically added to any Jira ticket that
is referenced in the pull-request/issue name. See wiki for more details.