Mention automatic JIRA linking as part of the "Get Involved" page
Mentions the capability to automatically referencing a pull request in a
JIRA ticket by including a reference to the ticket in the pull request
title in the "Get Involved" section.

The wording was chosen carefully not to imply that this would actually
link the state of the two. As far as I can tell, merging the pull
request will not have any influence on the state of the referenced
ticket. If this is actually the case, the wording should be adjusted to
reflect the behavior.

Moved the links to the JDO JIRA to HTTPS.