Use icons with highest resolution available for website
Adjusts the website to use the 150x150 version of the JDO logo instead
of smaller ones. Sets the scaling manually to keep the previous size
when using the larger image.

Adjusts the image size of the header icon to better fit the bar and
create an even spacing with the default layout.