Simplify workflow branch rule
Simplifies the deploy workflow branch rule by switching to specifying
which branches to run on.

I previously thought that 'branch' and 'path' rules were compared with
an OR operator. From my new tests, this does not seem to be the case.
Instead, they are compared with the AND operator, allowing for this
simplified configuration.